refactoring 2.
