+++ .gitignore
@@ -1,6 +1,6 @@
 *__pycache__*
 *.pyc
-data/*
+src/geoldm/data/*
 .idea*
 bash.sh
 kohler/cache/*
@@ -9,11 +9,11 @@ kohler/cache/*
 *.log
 *wandb*
 *.DS_Store*
-qm9/temp/*
+src/geoldm/qm9/temp/*
 kohler/qm9/temp/*
 push.sj
 bash_qm9.sh
-outputs/*
+src/geoldm/outputs/*
 bash_dw4.sh
 bash_qm9pos.sh
 bash_qm9_analyze.sh
@@ -26,7 +26,7 @@ dist/
 EN_diffusion.egg-info/
 */sidechainnet_data/*
 *example.pdb
-qm9/property_prediction/outputs/*
+src/geoldm/qm9/property_prediction/outputs/*
 qm9_conditional.sh
 figures/
 geom_drugs.sh
