+++ configs/paths.py
@@ -2,7 +2,4 @@ from pathlib import Path
 
 project_root = Path(__file__).parent.parent.absolute()
 
-ANALYSIS_DIR = project_root / "analysis"
-EGNN_DIR = project_root / "egnn"
-egnn_model = EGNN_DIR / "egnn"
-qm9 = project_root / "qm9"
+CROSSDOCKED_DATA = project_root / 'crossdocked'
