+++ geoldm.egg-info/requires.txt
@@ -1,0 +1,11 @@
+biopython>=1.85
+fair-esm>=2.0.0
+imageio>=2.37.0
+matplotlib>=3.10.3
+numpy>=2.2.5
+pip>=25.1.1
+rdkit>=2025.3.2
+scipy>=1.15.3
+torch>=2.7.0
+torch-geometric>=2.6.1
+wandb>=0.19.11
