+++ pyproject.toml
@@ -18,6 +18,14 @@ dependencies = [
     "wandb>=0.19.11",
 ]
 
+[build-system]
+requires = ["setuptools >= 61.0"]
+build-backend = "setuptools.build_meta"
+
+[tool.setuptools.packages.find]
+where = ["./"]
+include = ["src", "scripts"]
+
 
 [tool.ruff]
 line-length = 100
