+++ scripts/utils.py
@@ -6,7 +6,7 @@ import torch
 # Folders
 def create_folders(args):
     try:
-        os.makedirs('src/geoldm/outputs')
+        os.makedirs('../src/geoldm/outputs')
     except OSError:
         pass
