+++ src/geoldm/egnn/models.py
@@ -1,7 +1,7 @@
 import torch
 import torch.nn as nn
-from egnn.egnn_new import EGNN, GNN
-from equivariant_diffusion.utils import remove_mean, remove_mean_with_mask
+from src.geoldm.egnn.egnn_new import EGNN, GNN
+from src.geoldm.equivariant_diffusion.utils import remove_mean, remove_mean_with_mask
 import numpy as np
