--- src/geoldm/qm9/data/__init__.py
@@ -1,3 +1,0 @@
-from qm9.data.utils import initialize_datasets
-from qm9.data.collate import PreprocessQM9
-from qm9.data.dataset_class import ProcessedDataset
