--- src/geoldm/qm9/data/prepare/__init__.py
@@ -1,4 +1,0 @@
-from qm9.data.prepare.download import *
-from qm9.data.prepare.process import *
-from qm9.data.prepare.qm9 import *
-from qm9.data.prepare.md17 import *
