+++ src/geoldm/qm9/data/prepare/download.py
@@ -1,8 +1,8 @@
 import logging
 import os
 
-from qm9.data.prepare.md17 import download_dataset_md17
-from qm9.data.prepare.qm9 import download_dataset_qm9
+from src.geoldm.qm9.data.prepare.md17 import download_dataset_md17
+from src.geoldm.qm9.data.prepare.qm9 import download_dataset_qm9
 
 
 def prepare_dataset(datadir, dataset, subset=None, splits=None, cleanup=True, force_download=False):
