+++ src/geoldm/qm9/property_prediction/main_qm9_prop.py
@@ -1,6 +1,6 @@
 import sys
 import os
-sys.path.append(os.path.abspath(os.path.join('../../')))
+sys.path.append(os.path.abspath(os.path.join('../../../../')))
 from qm9.property_prediction.models_property import EGNN, Naive, NumNodes
 import torch
 from torch import nn, optim
