+++ src/geoldm/qm9/rdkit_functions.py
@@ -2,7 +2,7 @@ from rdkit import Chem
 from qm9.bond_analyze import get_bond_order, geom_predictor
 from . import dataset
 import torch
-from src.geoldm.configs import get_dataset_info
+from src.geoldm.configs.datasets_config import get_dataset_info
 import pickle
 import os
